requirements.txt
python-telegram-bot==20.7
opencv-python==4.8.1.78
moviepy==1.0.3
pillow==10.1.0
numpy==1.24.3